# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:           #раздел, имеющий 2 эндпоинта
    get:
      summary: Get all tasks    #что делает
      tags:                 #метка
        - tasks
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: Filter task by user ID
      responses:                #Что получаем
        '200':
          description: A list of tasks      #Описание
          content:                #Что лежит внутри
            application/json:
              schema:     #Структура данных
                type: array
                items:    #элемент массива
                  $ref: '#/components/schemas/Task' # Ссылка на схему Task
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:          #Тело (что нужно отправить)
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate' #структура данных
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    patch:
      summary: Update a task
      tags:
        - tasks
      parameters:
        - name: id    #параметр пути
          in: path    #параметр пути в url
          required: true
          schema:
            type: string  # строка
      requestBody:
        description: Task data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string 
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    patch:
      summary: Update a user
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody: 
        description: User data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    
    delete:
      summary: Delete a user
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
  /users/{id}/tasks:
    get:
      summary: Get all tasks for a specific user
      operationId: GetUsersIdTasks  # ← ВАЖНО ДОБАВИТЬ!
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      responses:
        '200':
          description: A list of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: User not found

components:   #Описание всех структурных данных
  schemas:
    Task:
      type: object
      properties:   #Свойства
        id:
          type: string  
        task:
          type: string
        is_done:
          type: string
        user_id:
          type: string
    TaskCreate:
      type: object
      required: [task, user_id]   #[] обозначает массив 
      properties:
        task:
          type: string
        is_done:
          type: string
        user_id:
          type: string
    TaskUpdate:
      type: object
      properties:
        task:
          type: string
        is_done:
          type: string
        user_id:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        tasks:                    # ← ДОБАВЛЯЕМ!
          type: array
          items:
            $ref: '#/components/schemas/Task'
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - email
        - created_at
        - updated_at
    UserCreate:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
        